requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
plotly==5.22.0
scikit-learn==1.2.2
tensorflow-cpu==2.16.2
requests-cache==1.2.1